Fix test errors when using checker testlib
Fix errors when attempting to run the tests using the gradle build
target. Checker framework testlib is not kept in sync with the main
framework so we cannot blindly update both to the same number. Stub
files are also not automatically loaded when using the gradle target
so we need to set them manually